+++ endpoints/users.py
@@ -1,7 +1,7 @@
 from app import app, db
 from helper.helpers import ModelEncoder
 from flask import request
-from models.models import Users
+from models.models import Users, Splits
 import json
 
 
@@ -34,3 +34,26 @@ def update_user_profile(id):
     user.progression_data = data.progression_data
     db.session.commit()
     return json.dumps(user.serialize(), cls=ModelEncoder)
+
+
+@app.route("/user/<id>/split", methods=['GET'])
+def get_user_splits(id):
+    data = []
+    splits = Splits.query.filter_by(user_id=id).all()
+    if not splits:
+        if Users.query.filter_by(discord_id=id).first() is None:
+            return "Could not find User", 404
+    for row in splits:
+        data.append(row.serialize())
+    return data
+
+
+@app.route("/user/<id>/split/total/", methods=['GET'])
+def get_user_total_splits(id):
+    if Users.query.filter_by(discord_id=id).first() is None:
+        return "Could not find User", 404
+    data = 0
+    splits = Splits.query.filter_by(user_id=id).all()
+    for row in splits:
+        data += row.split_amount
+    return str(data)
