+++ models/models.py
@@ -67,7 +67,8 @@ class Splits(db.Model, Serializer):
     id = db.Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     user_id = db.Column(db.String, db.ForeignKey('users.discord_id'))
     item_name = db.Column(db.String, nullable=False)
-    split_amount = db.Column(db.Numeric, nullable=False)
+    item_price = db.Column(db.Numeric, nullable=False)
+    split_contribution = db.Column(db.Numeric, nullable=False)
     group_size = db.Column(db.Integer, nullable=False)
     screenshot_link = db.Column(db.String)
     timestamp = db.Column(db.DateTime, nullable=False, default=datetime.datetime.now())
